+++ src/compute-engine/boxed-expression/numerics.ts
@@ -116,6 +116,9 @@ export function signDiff(
   const lhsN = lhs.N();
   const rhsN = rhs.N();
 
+  if (lhs.isNumberLiteral && lhs.im !== 0) return undefined;
+  if (rhs.isNumberLiteral && rhs.im !== 0) return undefined;
+
   // Structural equality?
   if (lhsN.isSame(rhsN)) return 0;