+++ test/playground.ts
@@ -2,6 +2,8 @@ import { ComputeEngine } from '../src/compute-engine';
 
 const ce = new ComputeEngine();
 
+console.log(ce.parse('2(13.1+x)<(10-5)').isEqual(ce.parse('26.2+2x<5')));
+
 // y powers should combine
 console.log(
   ce.parse('-2x5z\\sqrt{y}\\frac{3}{4}3\\pi y').simplify().toString()